Refactor I/O operation structures for platform-specific compatibility and memory management
- Updated `RawHandle` type to use platform-specific types for Unix and Windows, enhancing compatibility and clarity.
- Modified `Accept`, `Connect`, and related structs to utilize boxed payloads, improving memory efficiency and reducing struct size.
- Refactored `IocpOpPayload` and `UringOpPayload` to manage boxed payloads, ensuring consistent handling across I/O drivers.
- Enhanced lifecycle management functions for boxed payloads, improving safety and resource management in I/O operations.

+++ veloq-runtime/src/io/driver/uring.rs
@@ -104,7 +104,7 @@ impl UringDriver {
             return;
         }
 
-        let fd = self.waker_fd as usize;
+        let fd = self.waker_fd;
         let op = crate::io::op::Wakeup { fd: crate::io::op::IoFd::Raw(fd) };
         // Use into_platform_op to convert to UringOp
         let uring_op = <crate::io::op::Wakeup as IntoPlatformOp<UringDriver>>::into_platform_op(op);
